Speedgrader files reorganization (#3277)
