+++ Teacher/Teacher/SpeedGrader/MainLayout/View/SimilarityScore.swift
